Time: 34 ms (13.99%), Space: 11.4 MB (73.08%) - LeetHub

+++ 25-reverse-nodes-in-k-group/25-reverse-nodes-in-k-group.cpp
@@ -22,7 +22,7 @@ public:
     }
     ListNode* reverseKGroup(ListNode* head,const int k) {
         if(k==1) return head;
-        short int t;
+        int t;
         bool first = true;
         ListNode* temp = head,*ans,*prevstart=NULL,*tempans,*start;
         while(true){
